+++ src/Printers/ConfigurationPrinterInterface.ts
@@ -1,5 +1,7 @@
 import { ConfigurationItemInterface } from "../Schema/ConfigurationItemInterface";
-export interface PrinterInterface {
+
+export interface ConfigurationPrinterInterface {
     render(items: ConfigurationItemInterface[]): string;
+
     print(items: ConfigurationItemInterface[]): void;
 }