+++ src/Schema/ConfigurationBagInterface.ts
@@ -1,11 +1,11 @@
-import { PrinterInterface } from "../Printers/PrinterInterface";
+import { ConfigurationPrinterInterface } from "../Printers/ConfigurationPrinterInterface";
 
 export interface ConfigurationBagInterface {
     has(key: string): boolean;
 
     get(key: string): any;
 
-    render(printer: PrinterInterface): string;
+    render(printer: ConfigurationPrinterInterface): string;
 
-    print(printer: PrinterInterface): void;
+    print(printer: ConfigurationPrinterInterface): void;
 }