+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Rename `PrinterInterface` to `ConfigurationPrinterInterface`.
+
 ## [1.0.0] - 2019-10-17
 
 Initial version.